excel-agent-P3/excel-agent/requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.1

# AI/LLM Integration
google-generativeai>=0.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
python-dotenv>=1.0.0

# Optional: Enhanced visualization
plotly>=5.17.0

# Development/Testing
pytest>=7.0.0